feat: Refactor the way messages work
hx_messages was triggering an HXRequest which in turn sent
a trigger to the DOM which would then render the message.

Insetead the messages are leveraging hx-swap-oob and instead
is just being appended to the response html and using hx-swap-oob
it will be rendered into the DOM.

- Saves the need for an extra request
- Handles multiple messages now

This is a BREAKING CHANGE because the old way of rendering messages
won't work because there is no longer a trigger for the messages.

Additionally messages in context before was a 'message' and a 'tag'.
Now it's a list, 'messages', and each message has on it 'message.body'
and 'message.tag'.