feat: Change the way 'gets' are handled
This is a BREAKING CHANGE. Previously, the `get` method
was called by the `hx_requests` module, setting up the
necessarry attributes for the view's context. However,
that was causing performace issues and therefore that
has been replaced with a new method.

Django views' gets are still called by default (minus
the returning of the response which was the cause
of the performance issue), but any custom views that
were relying on the `get` method to be called will
now break.

To fix this, the get needs to be defined and
the HX_REQUESTS_CUSTOM_VIEWS_SETUP setting needs to be
set (see the docs for this)