fix!: change HX_REQUESTS_MODAL_BODY_SELECTOR to HX_REQUESTS_MODAL_BODY_ID
Using a class was not a good idea because there could be multiple
on the page.

BREAKING CHANGE: HX_REQUESTS_MODAL_BODY_SELECTOR is now
HX_REQUESTS_MODAL_BODY_ID and the default is now #hx_modal_body.
(set hx_modal_body as the html id on the modal body)