fix!: Use body_template instead of GET_template for HXModal
It was confusing because the GET_template is what was set which
would make you think that is what was being returned but in
reality it's returning the entire modal. By setting only the
body_template it's more clear that the body is being set and that
the request actually returns the entire modal.

BREAKING CHANGE: The GET_template is no longer used for the
HXModal. Instead the body_template is used. What was previously set
as GET_template should now be set as body_template.