feat!: add kwargs_as_context option to BaseHXRequest
Originally did not want to do this because was worried about
collisions with other context variables. However, after usage
it seems it would be more convenient to have the kwargs in the
context directly. Built in opt out though in case there are
collisions and then will be put in context like it is now
as hx_kwargs.

BREAKING CHANGE: kwargs will now be added to the context directly
by default. If you do not want this, you can set kwargs_as_context
to False. Any code that relies on kwargs being in the hx_kwargs
context variable will need to be updated to use the new context.
For example, if you were doing `hx_kwargs.some_key` in the
template you will need to change it to `some_key` OR set
kwargs_as_context to False.