feat!: Move view's get call to the hx_request
BREAKING CHANGE: Everything should work as before, however it
is breaking because the setup_hx_request function now has
parameters for *args and **kwargs. These are the view's args
and kwargs and are needed to setup the view from the hx_request

The reason for this change is because the old implementation does
not give room to override attributes of the view before the view's
get is called. Now by overriding the setup_view function in the hx_request
you can dynamically change view's attributes prior to the view's
get being called.