feat! Remove the need to setup views in get method
This was implemented because it was thought that it would
be bad for performance to call the get method of the view
but because of Django's TemplateResponses which all of their
built in views returns it is not a problem. Their TemplateView
is lazy rendered so nothing is evaluated unless it is rendered
which it is not because hx-requests returns other Html.

BREAKING CHANGE: The only real breaking change is that
hx_requests no longer have an extra_context attribute.
The removal of the custom gets should not be breaking because
the get method of the view is called and the context is
set up there. Technically if you did something different in the
custom views get than the actual vies get then it would break
that as well.