fix! Change modalFormValid to closeHxModal
It was misnamed as it really is just used to trigger a modal close event.

BREAKING CHANGE: This will break any custom modals that use
the modalFormValid event. They will need to be updated to use
closeHxModal instead. (and if using Alpin.js modal-form-valid
needs to be changed to close-hx-modal)