fix! Move setting of messages of out POST methods
This is because there may be other things you want to do
in form_valid that could determine the message that is set.
And they really belong in the form_valid and form_invalid methods
anyway.

BREAKING CHANGE: This will require you to set the messages in the
form_valid and form_invalid methods if you override them as opposed
to previously where the messages were set in the POST methods so they
were set even when overriding form_valid and form_invalid. The
DeleteHXRequest class has also been updated to reflect this change.
where messages are now set in the delete method and not the POST method.