feat: Add use_current_url option to BaseHxRequest
The idea is to replace use_full_path from the template tags because
it's finicky. It can sometimes cause double of same params in url and
it does not update because loaded on page load, so unless when you do
something like filter you ALSO replace the element that has the tag
with use_full_path, you are going to have a stale url for that request.

This use_current_url uses the HX-Current-URL header to get the current
URL which uses the browsers current url so is always up to date. When
set to true it combines the current URL's GET params with the GET params
of the request - but the ones from the request itself take precedence.

For now marked as Experimental because not fully tested. When confirmed,
docs will be added for it. And at some point use_full_path will be
removed due to it's flakiness.